renaming group id
