+++ sponge-api/src/main/java/io/github/destroyerofcode/common/Constants.java
@@ -1,0 +1,10 @@
+package io.github.destroyerofcode.common;
+
+public class Constants {
+  public static final int KECCAK_LANE = 5;
+  public static final int KECCAK_SIDE = KECCAK_LANE * KECCAK_LANE;
+
+  public static final int BITS_IN_LONG = 64;
+  public static final int BITS_IN_BYTE = 8;
+  public static final int BYTES_IN_LONG = BITS_IN_LONG / BITS_IN_BYTE; // 8
+}
