+++ sponge-api/src/main/java/io/github/destroyerofcode/exception/SpongeException.java
@@ -1,0 +1,12 @@
+package io.github.destroyerofcode.exception;
+
+public class SpongeException extends RuntimeException {
+
+  public SpongeException(final String message) {
+    super(message);
+  }
+
+  public SpongeException(final String message, final Throwable cause) {
+    super(message, cause);
+  }
+}
